+++ src/cdb/db.rs
@@ -1,11 +1,12 @@
 use {
     super::{
-        property, watch_cache::*, CertificationStore, Clause, ClauseDB, ClauseDBIF, ClauseId,
-        RefClause,
+        binary::{BinaryLinkIF, BinaryLinkList},
+        property,
+        watch_cache::*,
+        BinaryLinkDB, CertificationStore, Clause, ClauseDB, ClauseDBIF, ClauseId, RefClause,
     },
     crate::{assign::AssignIF, solver::SolverEvent, types::*},
     std::{
-        collections::HashMap,
         ops::{Index, IndexMut, Range, RangeFrom},
         slice::{Iter, IterMut},
     },
@@ -15,7 +16,7 @@ impl Default for ClauseDB {
     fn default() -> ClauseDB {
         ClauseDB {
             clause: Vec::new(),
-            bi_clause: Vec::new(),
+            binary_link: BinaryLinkDB::default(),
             watch_cache: Vec::new(),
             freelist: Vec::new(),
             certification_store: CertificationStore::default(),
@@ -159,15 +160,13 @@ impl Instantiate for ClauseDB {
         let nc = cnf.num_of_clauses;
         let mut clause = Vec::with_capacity(1 + nc);
         clause.push(Clause::default());
-        let mut bi_clause = Vec::with_capacity(2 * (nv + 1));
         let mut watcher = Vec::with_capacity(2 * (nv + 1));
         for _ in 0..2 * (nv + 1) {
-            bi_clause.push(BiClause::new());
             watcher.push(WatchCache::new());
         }
         ClauseDB {
             clause,
-            bi_clause,
+            binary_link: BinaryLinkDB::instantiate(config, cnf),
             watch_cache: watcher,
             certification_store: CertificationStore::instantiate(config, cnf),
             soft_limit: config.c_cls_lim,
@@ -212,11 +211,10 @@ impl Instantiate for ClauseDB {
                 }
             }
             SolverEvent::NewVar => {
+                self.binary_link.add_new_var();
                 // for negated literal
-                self.bi_clause.push(HashMap::new());
                 self.watch_cache.push(WatchCache::new());
                 // for positive literal
-                self.bi_clause.push(HashMap::new());
                 self.watch_cache.push(WatchCache::new());
                 self.lbd_temp.push(0);
             }
@@ -239,20 +237,13 @@ impl ClauseDBIF for ClauseDB {
         self.clause.iter_mut()
     }
     #[inline]
-    fn bi_clause_map(&self, l: Lit) -> &HashMap<Lit, ClauseId> {
-        &self.bi_clause[l]
+    fn binary_links(&self, l: Lit) -> &BinaryLinkList {
+        self.binary_link.connect_with(l)
     }
 
     // watch_cache_IF
     fn fetch_watch_cache_entry(&self, lit: Lit, wix: WatchCacheProxy) -> (ClauseId, Lit) {
-        #[cfg(feature = "hashed_watch_cache")]
-        {
-            todo!()
-        }
-        #[cfg(not(feature = "hashed_watch_cache"))]
-        {
-            self.watch_cache[lit][wix]
-        }
+        self.watch_cache[lit][wix]
     }
 
     #[inline]
@@ -266,8 +257,6 @@ impl ClauseDBIF for ClauseDB {
         self.watch_cache[l].swap_remove(iter.index);
         iter.detach_entry();
     }
-    fn reregister_watch_cache(&mut self, _: Lit, _: Option<WatchCacheProxy>) {}
-    fn restore_detached_watch_cache(&mut self, _: Lit, _: WatchCacheIterator) {}
     fn merge_watch_cache(&mut self, p: Lit, wc: WatchCache) {
         self.watch_cache[p].append_watch(wc);
     }
@@ -335,7 +324,7 @@ impl ClauseDBIF for ClauseDB {
             ref mut num_bi_learnt,
             ref mut num_lbd2,
             ref mut num_learnt,
-            ref mut bi_clause,
+            ref mut binary_link,
             ref ordinal,
             ref mut watch_cache,
             ..
@@ -381,8 +370,7 @@ impl ClauseDBIF for ClauseDB {
         if len2 {
             // assert_eq!(c.lits.len(), 2);
             *num_bi_clause += 1;
-            bi_clause[l0].insert(l1, cid);
-            bi_clause[l1].insert(l0, cid);
+            binary_link.add(l0, l1, cid);
         } else {
             // assert!(2 < c.lits.len());
             watch_cache[!l0].insert_watch(cid, l1);
@@ -418,7 +406,7 @@ impl ClauseDBIF for ClauseDB {
         let ClauseDB {
             ref mut clause,
             ref mut lbd_temp,
-            ref mut bi_clause,
+            ref mut binary_link,
             ref ordinal,
             ref mut watch_cache,
             ..
@@ -442,8 +430,7 @@ impl ClauseDBIF for ClauseDB {
         let l0 = c.lits[0];
         let l1 = c.lits[1];
         if len2 {
-            bi_clause[l0].insert(l1, cid);
-            bi_clause[l1].insert(l0, cid);
+            binary_link.add(l0, l1, cid);
         } else {
             watch_cache[!l0].insert_watch(cid, l1);
             watch_cache[!l1].insert_watch(cid, l0);
@@ -461,7 +448,7 @@ impl ClauseDBIF for ClauseDB {
         debug_assert!(1 < c.lits.len());
         remove_clause_fn(
             &mut self.certification_store,
-            &mut self.bi_clause,
+            &mut self.binary_link,
             &mut self.watch_cache,
             &mut self.num_bi_clause,
             &mut self.num_clause,
@@ -482,7 +469,7 @@ impl ClauseDBIF for ClauseDB {
         let mut dummy3 = 1;
         remove_clause_fn(
             &mut store,
-            &mut self.bi_clause,
+            &mut self.binary_link,
             &mut self.watch_cache,
             &mut dummy1,
             &mut dummy2,
@@ -499,14 +486,8 @@ impl ClauseDBIF for ClauseDB {
         iter: &mut WatchCacheIterator,
         op: Option<Lit>,
     ) {
-        #[cfg(feature = "hashed_watch_cache")]
-        todo!();
-
-        #[cfg(not(feature = "hashed_watch_cache"))]
-        {
-            if let Some(p) = op {
-                self.watch_cache[l][iter.index].1 = p;
-            }
+        if let Some(p) = op {
+            self.watch_cache[l][iter.index].1 = p;
         }
 
         // let cid = self.watch_cache[l][iter.index].0;
@@ -535,7 +516,7 @@ impl ClauseDBIF for ClauseDB {
         debug_assert!(1 < self[cid].len());
         let ClauseDB {
             ref mut clause,
-            ref mut bi_clause,
+            ref mut binary_link,
             ref mut watch_cache,
             ref mut certification_store,
             ref mut num_bi_clause,
@@ -564,11 +545,11 @@ impl ClauseDBIF for ClauseDB {
             .copied()
             .collect::<Vec<Lit>>();
         if new_lits.len() == 2 {
-            if let Some(reg) = bi_clause[new_lits[0]].get(&new_lits[1]) {
+            if let Some(reg) = binary_link.search(new_lits[0], new_lits[1]) {
                 //
                 //## Case:3-0
                 //
-                return RefClause::RegisteredClause(*reg);
+                return RefClause::RegisteredClause(reg);
             }
             //
             //## Case:3-2
@@ -578,8 +559,7 @@ impl ClauseDBIF for ClauseDB {
             std::mem::swap(lits, &mut new_lits);
             watch_cache[!l0].remove_watch(&cid);
             watch_cache[!l1].remove_watch(&cid);
-            bi_clause[lits[0]].insert(lits[1], cid);
-            bi_clause[lits[1]].insert(lits[0], cid);
+            binary_link.add(lits[0], lits[1], cid);
             *num_bi_clause += 1;
             // self.watches(cid, "after strengthen_by_elimination case:3-2");
         } else {
@@ -673,7 +653,7 @@ impl ClauseDBIF for ClauseDB {
         debug_assert!(!self[cid].is_dead());
         let ClauseDB {
             clause,
-            bi_clause,
+            binary_link,
             watch_cache,
             certification_store,
             ..
@@ -681,8 +661,7 @@ impl ClauseDBIF for ClauseDB {
         let c = &mut clause[std::num::NonZeroU32::get(cid.ordinal) as usize];
         debug_assert!(new_lits.len() < c.len());
         if new_lits.len() == 2 {
-            if let Some(bc) = bi_clause[new_lits[0]].get(&new_lits[1]) {
-                let did = *bc;
+            if let Some(did) = binary_link.search(new_lits[0], new_lits[1]) {
                 //
                 //## Case:0
                 //
@@ -701,8 +680,7 @@ impl ClauseDBIF for ClauseDB {
             let l1 = c.lit0();
             watch_cache[!old_l0].remove_watch(&cid);
             watch_cache[!old_l1].remove_watch(&cid);
-            bi_clause[l0].insert(l1, cid);
-            bi_clause[l1].insert(l0, cid);
+            binary_link.add(l0, l1, cid);
 
             if certification_store.is_active() {
                 certification_store.add_clause(new_lits);
@@ -807,7 +785,7 @@ impl ClauseDBIF for ClauseDB {
         }
         let ClauseDB {
             clause,
-            bi_clause,
+            binary_link,
             watch_cache,
             certification_store,
             ..
@@ -827,8 +805,7 @@ impl ClauseDBIF for ClauseDB {
                 let l0 = new_lits[0];
                 let l1 = new_lits[1];
                 debug_assert!(2 < c.lits.len());
-                if let Some(r) = bi_clause[l0].get(&l1) {
-                    let bid = *r;
+                if let Some(bid) = binary_link.search(l0, l1) {
                     //
                     //## Case:3-0
                     //
@@ -840,8 +817,7 @@ impl ClauseDBIF for ClauseDB {
                 //
                 watch_cache[!c.lits[0]].remove_watch(&cid);
                 watch_cache[!c.lits[1]].remove_watch(&cid);
-                bi_clause[l0].insert(l1, cid);
-                bi_clause[l1].insert(l0, cid);
+                binary_link.add(l0, l1, cid);
                 std::mem::swap(&mut c.lits, &mut new_lits);
                 self.num_bi_clause += 1;
                 c.turn_off(FlagClause::LEARNT);
@@ -1086,7 +1062,7 @@ impl ClauseDBIF for ClauseDB {
             {
                 remove_clause_fn(
                     &mut self.certification_store,
-                    &mut self.bi_clause,
+                    &mut self.binary_link,
                     &mut self.watch_cache,
                     &mut self.num_bi_clause,
                     &mut self.num_clause,
@@ -1135,7 +1111,7 @@ impl ClauseDBIF for ClauseDB {
         }
         let l0 = vec[0];
         let mut num_sat = 0;
-        for (_, &cid) in self.bi_clause[l0].iter() {
+        for (_, cid) in self.binary_link.connect_with(l0).iter() {
             let c = &self.clause[std::num::NonZeroU32::get(cid.ordinal) as usize];
             debug_assert!(c[0] == l0 || c[1] == l0);
             let other = c[(c[0] == l0) as usize];
@@ -1307,10 +1283,10 @@ impl ClauseDB {
     fn complete_bi_clauses_with(&mut self, asg: &mut impl AssignIF, lit: Lit) {
         let mut vec: Vec<Vec<Lit>> = Vec::new();
         // [lit, other]
-        for other in self.bi_clause[lit].keys() {
+        for (other, _) in self.binary_link.connect_with(lit) {
             // [!other, third]
-            for third in self.bi_clause[!*other].keys() {
-                if lit.vi() != third.vi() && !self.bi_clause[lit].contains_key(third) {
+            for (third, _) in self.binary_link.connect_with(!*other) {
+                if lit.vi() != third.vi() && self.binary_link.search(lit, *third).is_none() {
                     // the new [lit, third] should be added.
                     vec.push(vec![lit, *third]);
                 }
@@ -1349,7 +1325,7 @@ impl ClauseDB {
     /// bi_clause[l0].get(&l1).is_some()
     ///```
     fn has_bi_clause(&self, l0: Lit, l1: Lit) -> Option<ClauseId> {
-        self.bi_clause[l0].get(&l1).copied()
+        self.binary_link.search(l0, l1)
     }
     #[cfg(feature = "strategy_adaptation")]
     /// change good learnt clauses to permanent one.
@@ -1393,7 +1369,7 @@ impl ClauseDB {
 #[allow(clippy::too_many_arguments)]
 fn remove_clause_fn(
     certification_store: &mut CertificationStore,
-    bi_clause: &mut [BiClause],
+    binary_link: &mut BinaryLinkDB,
     watcher: &mut [WatchCache],
     num_bi_clause: &mut usize,
     num_clause: &mut usize,
@@ -1405,8 +1381,9 @@ fn remove_clause_fn(
     let l0 = c.lits[0];
     let l1 = c.lits[1];
     if c.len() == 2 {
-        bi_clause[usize::from(l0)].remove(&l1); // .expect("db1072");
-        bi_clause[usize::from(l1)].remove(&l0); // .expect("db1073");
+        binary_link
+            .remove(l0, l1)
+            .expect("Eror (remove_clause_fn#01)");
         *num_bi_clause -= 1;
     } else {
         watcher[usize::from(!l0)].remove_watch(&cid); // .expect("db1076");